ProjMgtService-Refactoring: Refactoring the package for entities.
