+++ src/main/java/com/umssonline/proymgt/controllers/BacklogRestController.java
@@ -1,8 +1,8 @@
 package com.umssonline.proymgt.controllers;
 
 import com.umssonline.proymgt.models.dto.BacklogDto;
-import com.umssonline.proymgt.models.Backlog;
-import com.umssonline.proymgt.models.UserStory;
+import com.umssonline.proymgt.models.entity.Backlog;
+import com.umssonline.proymgt.models.entity.UserStory;
 import com.umssonline.proymgt.services.BacklogService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
