+++ src/main/java/com/umssonline/proymgt/controllers/ProjectRestController.java
@@ -1,8 +1,8 @@
 package com.umssonline.proymgt.controllers;
 
-import com.umssonline.proymgt.models.Backlog;
-import com.umssonline.proymgt.models.Project;
-import com.umssonline.proymgt.models.Sprint;
+import com.umssonline.proymgt.models.entity.Backlog;
+import com.umssonline.proymgt.models.entity.Project;
+import com.umssonline.proymgt.models.entity.Sprint;
 import com.umssonline.proymgt.services.ProjectService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
