+++ src/main/java/com/umssonline/proymgt/models/entity/Status.java
@@ -1,4 +1,4 @@
-package com.umssonline.proymgt.models;
+package com.umssonline.proymgt.models.entity;
 
 public enum Status {
     TO_BE_DISCUSED,
