+++ src/main/java/com/umssonline/proymgt/models/entity/UserType.java
@@ -1,4 +1,4 @@
-package com.umssonline.proymgt.models;
+package com.umssonline.proymgt.models.entity;
 
 public enum UserType {
     PRODUCT_OWNER,
