+++ src/main/java/com/umssonline/proymgt/repositories/ProjectRepository.java
@@ -1,6 +1,6 @@
 package com.umssonline.proymgt.repositories;
 
-import com.umssonline.proymgt.models.Project;
+import com.umssonline.proymgt.models.entity.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
