+++ src/main/java/com/umssonline/proymgt/repositories/SprintRepository.java
@@ -1,11 +1,10 @@
 package com.umssonline.proymgt.repositories;
 
-import com.umssonline.proymgt.models.Sprint;
+import com.umssonline.proymgt.models.entity.Sprint;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Collection;
-import java.util.Optional;
 
 @Repository
 public interface SprintRepository extends JpaRepository<Sprint, Long> {
