+++ src/main/java/com/umssonline/proymgt/repositories/TaskRepository.java
@@ -1,6 +1,6 @@
 package com.umssonline.proymgt.repositories;
 
-import com.umssonline.proymgt.models.Task;
+import com.umssonline.proymgt.models.entity.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
