+++ src/main/java/com/umssonline/proymgt/repositories/UserRepository.java
@@ -1,6 +1,6 @@
 package com.umssonline.proymgt.repositories;
 
-import com.umssonline.proymgt.models.User;
+import com.umssonline.proymgt.models.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
