+++ src/main/java/com/umssonline/proymgt/repositories/UserStoryRepository.java
@@ -1,6 +1,6 @@
 package com.umssonline.proymgt.repositories;
 
-import com.umssonline.proymgt.models.UserStory;
+import com.umssonline.proymgt.models.entity.UserStory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
